+++ Readme/CHANGELOG.md
@@ -32,6 +32,7 @@ If you would like to see fresh changes done since last release you can do so on 
 - Modules
 
   - Updated `Test-ExecutableFile` so that digital signature mismatch is always reported
+  - Fixed a bug in `Edit-Table` which didn't account for paths with environment variables
 
 - Scripts